search_perf - finish rewriting search function
