+++ src/main.rs
@@ -107,7 +107,7 @@ mod api_tests {
         let client = Client::tracked(rocket()).expect("Valid Rocket Instance");
         let res = client.get(uri!("/api/version")).dispatch();
         assert_eq!(res.status(), Status::Ok);
-        assert_eq!(res.into_string().unwrap(), r#"{"version":"2.5.1"}"#);
+        assert_eq!(res.into_string().unwrap(), r#"{"version":"2.5.2"}"#);
         cleanup();
     }
