+++ src/model/error/file_errors.rs
@@ -1,3 +1,5 @@
+use std::fmt::{Display, Formatter};
+
 #[derive(PartialEq, Debug)]
 pub enum CreateFileError {
     FailWriteDisk,
@@ -47,4 +49,13 @@ pub enum SearchFileError {
     DbError,
     /// an issue occurred retrieving tags
     TagError,
+}
+
+impl Display for SearchFileError {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::DbError => write!(f, "SearchFileError::DbError"),
+            Self::TagError => write!(f, "SearchFileError::TagError"),
+        }
+    }
 }
