+++ src/service/mod.rs
@@ -1,4 +1,5 @@
 pub mod api_service;
 pub mod file_service;
 pub mod folder_service;
+pub mod search_service;
 pub mod tag_service;
